Fix: Remove explicit JDK versioning from `SystemUtil`
As it stands, our SDK only supports three old LTS versions of Java.
Any user who wants to use the SDK on non-LTS versions, or on
reasonably new versions of Java, will be prevented.  However, what the
code really wants is to know whether we’re on a version of Java before
or after Java 9, with which we determine which CRC32 implementation to
use.

This patch *removes* explicit JDK versioning from `SystemUtil`, which
is a breaking change.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>