Fix: avoid implicit cast from `long` to `int`
Compiling on modern JDKs results in the following warning:

    [WARNING] blazingmq-sdk-java/bmq-sdk/src/main/java/com/bloomberg/bmq/impl/infr/proto/MessagePropertiesImpl.java:[234,24] implicit cast from long to int in compound assignment is possibly lossy

This is fired on the line:

    // Add padding bytes
    totalLength += numPaddingBytes;

where `totalLength` is an `int` and `numPaddingBytes` is a `long`.
However, `numPaddingBytes` is initialized few lines above as

    // Read padding bytes
    final long numPaddingBytes = input.readByte();

Since `readByte()` returns a `byte`, there’s no reason for this
constant to be `long`.  This patch fixes the above warning by avoiding
the needless conversion to `long`.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>